requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
arch==8.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
cachetools==6.2.4
catboost==1.2.8
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cloudpickle==3.1.2
cmdstanpy==1.3.0
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
Cython==3.2.3
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.1
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
fsspec==2025.12.0
gast==0.7.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
graphviz==0.21
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
holidays==0.87
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imbalanced-learn==0.13.0
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==8.37.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
keras==2.15.0
kiwisolver==1.4.9
lark==1.3.1
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.6.0
lime==0.2.0.1
llvmlite==0.46.0
Mako==1.3.10
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.2.0
ml-dtypes==0.2.0
mpmath==1.3.0
narwhals==2.14.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.5.1
notebook_shim==0.2.4
numba==0.63.1
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optuna==4.6.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
patsy==1.0.2
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
pmdarima==2.0.4
prometheus_client==0.23.1
prompt_toolkit==3.0.52
prophet==1.2.1
protobuf==4.25.8
psutil==7.2.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.3.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.3
shap==0.49.1
simsimd==6.5.12
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.8
soupsieve==2.8.1
SQLAlchemy==2.0.45
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.6
stringzilla==4.6.0
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-addons==0.22.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
tomli==2.3.0
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typeguard==2.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.2
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.4
wrapt==1.14.2
xgboost==3.1.2
